fix(candidat-individuel): serialize quote publication

+++ prisma/migrations/20260824090000_add_profil_candidat/migration.sql
@@ -23,7 +23,7 @@ CREATE TABLE "profils_candidats" (
     "estRedoublant" BOOLEAN NOT NULL DEFAULT false,
     "estTitulaireBacDejaObtenu" BOOLEAN NOT NULL DEFAULT false,
     "brancheBascule" "BrancheBascule",
-    "moyenneRattrapage" INTEGER,
+    "moyenneRattrapage" DOUBLE PRECISION,
     "optionsTerminale" TEXT[] DEFAULT ARRAY[]::TEXT[],
     "notesConservees" JSONB,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
